Refactor Unit Tests
